Trying to get the image to load.
I've tried multiple different ways to load the image and none have worked. simply leaving it as the file name like I changed it too got it to load in the rStudio window. But, still didn't display once knitted. I believe it's an issue with the shinyApp. Here was a link that may help: https://stackoverflow.com/questions/21996887/embedding-image-in-shiny-app